Add files via upload
Minor changes to the Deck- so you can't try and pull from an invalid index, The old player (attack, reinforce & fortify) are commented out in the player class & the new ones are now in PlayerType. I added the map test files that I use. Initializing the armies is now automatic so you don't have to waste time. Also every player defends an attack automatically however there is a function in Player that gives an option.